Merged PR 4299: Revert 'Disabling the FRL pipeline'
<!--
#csf:3.3.5h:We use a PR template to remind everyone of our way of working and the Cloud Security Framework.
-->
## Why this PR is needed:

According to Niels, it will fail a few times and it will catch up with the download.

The reason for errors:

Multiple zip files are downloaded by the pipeline. As soon as the first zip file is downloaded, the 17 pipelines which convert CSV to parquet trigger which makes the queue of the ZIP#2 download queue for a long time.

After an hour, the ZIP#2 should be downloaded and again, 17 pipeline runs will start...

This will continue until all files are downloaded at which point nothing should happen anymore