Merged PR 4386: CDP-5021/split release of production triggers and azure function
<!--
#csf:3.3.5h:We use a PR template to remind everyone of our way of working and the Cloud Security Framework.
-->
## Why this PR is needed:

Due to the mismatch of the versions with python in the function code and function app, we are adding the hardcoded python version to the code - note that this has already been done for the production environment - https://dev.azure.com/NIBC/Data%20Platform/_git/adf-triggers-dtap/pullrequest/4339.

Additionally, we're splitting the release of the triggers and the function code due to the latter being changed much less often and possibly causing issues because of devops changing it's mind on something.

NOTE: we need to be able to re-start the release of the function without code change - for example for when we change the function app. Due to this, i decided to keep the condition to just the new tag.

## Related Tickets & Documents

- link to Jira issue to be solved by this PR:
  - https://jira/browse/CDP-5021

## I (creator of this PR):

- [ ] ✨ have chosen a clear title for this PR
  - This will greatly help my colleagues understand where changes come from
  - Template for PR name: starting with the related Jira issue and actively describing what this change does, e.g. `CDP-2024/Add PR Template`
- [ ] 👷‍♀️ created a small PR covering a single issue, allowing for effective review.
- [ ] 🙌 am aware of and followed best practices documented on [the wiki](https://dev.azure.com/NIBC/Data%20Platform/_wiki/wikis/Data-Platform.wiki/1525/Azure-Data-Factory)
- [ ] 💫 wrote clear, explicit code and added short comments where needed
- [ ] 📗 updated any related documentation and included any relevant screenshots.
- [ ] 🔐 am aware of the [cloud security framework](https://dev.azure.com/NIBC/Data%20Platform/_git/cloud-security-framework-evidence) and this PR does not go against the measure we take to be in control of our security