Merged PR 5562: CDP-6034/Removing all of the DEV triggers which use Cloud Integration Platform
<!--
#csf:3.3.5h:We use a PR template to remind everyone of our way of working and the Cloud Security Framework.
-->
## Why this PR is needed:

After internal discussions and suggestions from the CIP team, we've decided to not schedule any pipelines which rely on CIP on development environment.

The `https://apim.dev.cloud.nibc.com/` is used for internal development of the CIP team and is not stable environment. This means every so often we get a number of errors which cause us to spent unneeded time investigating issues which are not related to our projects.

Additional consideration for removal of DEV CIP triggers - it was discussed with the architect that we should not schedule our dev environment. This is a first step in that direction.

## Related Tickets & Documents

- link to Jira issue to be solved by this PR:
  - https://jira/browse/CDP-6034

## Example outcome of this change

Full removal of all pipelines using CIP on dev env

## Instructions on how to verify quality of this PR

Could you double check after me that I'm not missing any of the pipelines? I think i got them all but it's good to double check.

## [optional] Are there any post deployment tasks we need to perform?

No, this will only affect the development environment.

## I (creator of this PR):

- [X] ✨ have chosen a clear title for this PR
  - This will greatly help my colleagues understand where changes come from
  - Template for PR name: starting with the related Jira issue and actively describing what this change does, e.g. `CDP-2024/Add PR Template`
- [X] 👷‍♀️ created a small PR covering a single issue, allowing for effective review.
- [X] 🙌 am aware of and followed best practices documented on [the wiki](https://dev.azure.com/NIBC/Data%20Platform/_wiki/wikis/Data-Platform.wiki/1525/Azure-Data-Factory)
- [X] 💫 wrote clear, explicit code and added short comments where needed
- [X] 📗 updated any related documentation and included any relevant screenshots.
- [X] 🔐 am aware of the [cloud security framework](https://dev.azure.com/NIBC/Data%20Platform/_git/cloud-security-framework-evidence) and this PR does not go against the measure we take to be in control of our security